Fix indentation error
Needs to be inside both for/if-then statements.

Being outside the for loop means only the last site is being sent to
redcap.

Being outside the if-then statement means that there is the
potential the variable will not be declared, but still used. Best case
scenario this throws an error on first loop. Worst case scenario, this
silently uses the value from prior iteration on subsequent loops.